% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{checkboxMultiGroupInput}
\alias{checkboxMultiGroupInput}
\title{A group of checkboxes divided by sub-labels}
\usage{
checkboxMultiGroupInput(variable, choicesByLabel = NULL, selected = NULL)
}
\arguments{
\item{variable}{The name of the input group}

\item{choicesByLabel}{The choices indexed by label}

\item{selected}{Name of the query to plot}
}
\description{
A group of checkboxes divided by sub-labels
}
