% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getSubcategoryNames}
\alias{getSubcategoryNames}
\title{Get subcategory names}
\usage{
getSubcategoryNames(subcategories)
}
\arguments{
\item{subcategories}{a list of subcategories}
}
\description{
Get subcategory names
}
