% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{loadDefault}
\alias{loadDefault}
\title{Load default data into UI}
\usage{
loadDefault(regionSettings)
}
\arguments{
\item{regionsSettings}{the region to region group mappings}
}
\description{
Returns the data from the default data file
}
