% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{loadDefaultGroupColors}
\alias{loadDefaultGroupColors}
\title{Load the default sector colors}
\usage{
loadDefaultGroupColors()
}
\description{
Returns the sector colors from the default project file
}
