% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{loadDefaultPercentileColors}
\alias{loadDefaultPercentileColors}
\title{Load the default percentile colors}
\usage{
loadDefaultPercentileColors()
}
\description{
Returns the percentile colors from the default project file
}
