% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{loadDefaultRegionSettings}
\alias{loadDefaultRegionSettings}
\title{Load the default region colors}
\usage{
loadDefaultRegionSettings()
}
\description{
Returns the region colors from the default project file
}
